+++ NEWS.md
@@ -4,6 +4,10 @@
 
 * `ggplotly()` now supports the `{ggridges}` package. (#2314)
 
+## Improvements
+
+* `ggplotly()` now works better with the development version of ggplot2 (> v3.4.4). (#2315)
+
 ## Bug fixes
 
 * Closed #2337: Creating a new `event_data()` handler no longer causes a spurious reactive update of existing `event_data()`s. (#2339)